meet_1/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
gunicorn==22.0.0
Pillow==10.3.0
numpy==1.26.4
onnxruntime==1.19.2